refactor: move function to disruptions context
